Release 3.12.2: - API service as Android foreground service - Message parsing refactored to CarData, exception encapsulation - CarData: added new fields (SOH, drive mode & profile info, power, energy usage) - CarData: changed numeric integer fields to float/double for OVMS v3 - GcmListener: added notification type (v3 info/alert distinction) - UI: rearranged main tabs, allow all tabs in cache - UI: added help & monospace font options menu to notifications tab - UI: added options menu to car tab (direct access to info/features/diag logs) - Database: fixed "sticky" notifications (wrong future timestamp) - Features editor v3 support: changed to text editor, allow variable number of features, added new Twizy features - General crash hardening - Notifications: added sending modem ("@…") & msg commands ("#…") - UI: options menu optimizations

+++ OpenVehicleApp/src/main/java/com/openvehicles/OVMS/entities/GPSLogData.java
@@ -28,7 +28,7 @@ public class GPSLogData {
 
 		public float odometerMi;
 		public double latitude, longitude;
-		public int altitude, direction, speed;
+		public float altitude, direction, speed;
 		public int gspFix, gpsStaleCnt, gsmSignal;
 
 		public int currentPower, powerUsedWh, powerRecdWh;
@@ -322,9 +322,9 @@ public class GPSLogData {
 							entry.odometerMi = Integer.parseInt(result[j++]) / 10f;
 							entry.latitude = Double.parseDouble(result[j++]);
 							entry.longitude = Double.parseDouble(result[j++]);
-							entry.altitude = Integer.parseInt(result[j++]);
-							entry.direction = Integer.parseInt(result[j++]);
-							entry.speed = Integer.parseInt(result[j++]);
+							entry.altitude = Float.parseFloat(result[j++]);
+							entry.direction = Float.parseFloat(result[j++]);
+							entry.speed = Float.parseFloat(result[j++]);
 							entry.gspFix = Integer.parseInt(result[j++]);
 							entry.gpsStaleCnt = Integer.parseInt(result[j++]);
 							entry.gsmSignal = Integer.parseInt(result[j++]);
